remove a bunch of uncessary casts and other warnings from eclipse
git-svn-id: http://red5.googlecode.com/svn/java/server/trunk@3688 1b6495e4-3631-0410-8e05-8f51eee8b9cc

+++ test/org/red5/server/service/ConversionUtilsTest.java
@@ -91,7 +91,7 @@ public class ConversionUtilsTest {
 		int[] results = (int[]) result;
 
 		assertEquals(results.length, source.size());
-		assertEquals(results[2], ((Integer)source.get(2)).intValue());
+		assertEquals(results[2], source.get(2).intValue());
 	}
 	
 	@Test
